pyproject.toml
[project]
name = "agent-prod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.123.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "rocketmq-python-client==5.0.6",
    "uvicorn>=0.38.0",
]
